refactor: 메일 서비스 사용 시 회원 도메인, 인증 도메인 요청 구분
